+++ OJGIR/Mesh.h
@@ -28,9 +28,9 @@ public:
 	Mesh();
 	~Mesh();
 
-	glm::vec3 getPosition(){ return position; };
-	glm::mat4 getOrientation(){ return orientation; };
-	glm::mat4 getmodelMat(){ return modelMat; };
+	glm::dvec3 getPosition(){ return position; };
+	glm::dmat4 getOrientation(){ return orientation; };
+	glm::dmat4 getmodelMat(){ return modelMat; };
 	void getPosition(glm::vec3* vec){ vec->x = position[0]; vec->y = position[1]; vec->z = position[2]; };
 	void getOrientation(glm::mat4* mat);
 	
@@ -40,9 +40,9 @@ public:
 																							}
 	void setOrientation(glm::mat4 o) {orientation = o;// std::copy(o, o + 16, orientation); 
 																							}
-	glm::vec3 BRDF() { return brdf; }
-	float getP() { return P; }
-	glm::vec3 getLightEmission() { return lightEmission; };
+	glm::dvec3 BRDF() { return brdf; }
+	double getP() { return P; }
+	glm::dvec3 getLightEmission() { return lightEmission; };
 
 	vertex* getVarray(){ return vertexArray; }
 	triangle* getTarray(){ return indexArray; }
@@ -63,18 +63,18 @@ protected:
 	triangle* indexArray;
 	int nTris;
 
-	glm::vec3 position;
-	glm::mat4 orientation;
+	glm::dvec3 position;
+	glm::dmat4 orientation;
 
-	glm::mat4 modelMat;
+	glm::dmat4 modelMat;
 
 	GLuint vao;          // Vertex array object, the main handle for geometry
 	GLuint vertexbuffer; // Buffer ID to bind to GL_ARRAY_BUFFER
 	GLuint indexbuffer;  // Buffer ID to bind to GL_ELEMENT_ARRAY_BUFFER
 
 	//shading variables
-	glm::vec3 lightEmission;
-	glm::vec3 brdf;
-	float P;
+	glm::dvec3 lightEmission;
+	glm::dvec3 brdf;
+	double P;
 };
